pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oats-ai"
version = "0.6.0"
description = "OATS — On-demand Agent Tool Synthesis. LLMs synthesize and execute one-shot tools with mandatory human approval."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Trent", email = "trent@agenticwork.io" }
]
keywords = [
    "llm",
    "ai",
    "agents",
    "tools",
    "mcp",
    "agentic",
    "one-shot",
    "dynamic-tools",
    "synthesis",
    "oats",
    "tool-synthesis"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
]
embeddings = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
all = [
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[project.scripts]
oats = "oats.cli.main:app"

[project.urls]
Homepage = "https://oats.agenticwork.io"
Documentation = "https://oats-docs.agenticwork.io"
Repository = "https://github.com/agentic-work/oats"

[tool.hatch.build.targets.wheel]
packages = ["oats"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

# Enable useful checks without full strict mode
check_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_return_any = false
no_implicit_reexport = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
